refactor(app_runtime): Window handler を window.rs に抽出 (SPEC-2077 Phase C)
- crates/gwt/src/app_runtime/window.rs を新規作成し、 11 個の window handler を移行: create_window_events / focus_window_events / cycle_focus_events / update_viewport_events / arrange_windows_events / maximize_window_events / minimize_window_events / restore_window_events / update_window_geometry_events / close_window_events / list_windows_event
- mod.rs に mod window; 宣言を追加 (board / memory の隣)
- god module 解体 phase chain Phase C、 Phase A/B pattern を reuse

Verification:
- cargo test -p gwt-core -p gwt: 既知 flake (viewport_and_geometry_updates_persist_workspace_state) 1 件以外全 pass、 個別実行で pass 確認
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
- mod.rs 7190→6982 行 (Phase C で 208 行減、 累計 463 行減)
- window.rs 新 240 行 (window handler 単一責務)
